+++ api/src/hyperion/dev/spec/types.clj
@@ -1,0 +1,7 @@
+(ns hyperion.dev.spec.types
+  (:require [speclj.core :refer :all]
+            [hyperion.dev.spec.types.boolean :refer [it-handles-booleans]]))
+
+(defn it-handles-types []
+  (context "booleans"
+    (it-handles-booleans)))