+++ postgres/spec/hyperion/postgres_spec.clj
@@ -16,7 +16,7 @@
     (make-query query)))
 
 (def create-table-query
-  "CREATE TABLE %s (
+  "CREATE TABLE IF NOT EXISTS %s (
     id SERIAL PRIMARY KEY,
     name VARCHAR(35),
     first_name VARCHAR(35),
@@ -26,18 +26,27 @@
 
 (defn create-key-tables []
   (do-query
-    "CREATE TABLE account (
+    "CREATE TABLE IF NOT EXISTS account (
     id SERIAL PRIMARY KEY,
     first_name VARCHAR(35),
     inti INTEGER,
     data VARCHAR(32)
     );
-    CREATE TABLE shirt (
+    CREATE TABLE IF NOT EXISTS shirt (
     id SERIAL PRIMARY KEY,
     account_id INTEGER REFERENCES account,
     first_name VARCHAR(35),
     inti INTEGER,
     data VARCHAR(32)
+    )"))
+
+(defn create-types-table []
+  (do-query
+    "CREATE TABLE IF NOT EXISTS types (
+    id SERIAL PRIMARY KEY,
+    inti INTEGER,
+    data VARCHAR(32),
+    bool BOOLEAN
     )"))
 
 (defn create-table [table-name]
@@ -57,14 +66,16 @@
     (with-connection connection-url
       (create-table "testing")
       (create-table "other_testing")
-      (create-key-tables)))
+      (create-key-tables)
+      (create-types-table)))
 
   (after
     (with-connection connection-url
       (drop-table "testing")
       (drop-table "other_testing")
       (drop-table "shirt")
-      (drop-table "account")))
+      (drop-table "account")
+      (drop-table "types")))
 
   (it-behaves-like-a-datastore)
 
@@ -86,4 +97,5 @@
 
   (context "Transactions"
     (include-transaction-specs connection-url))
+
   )