+++ sqlite/spec/hyperion/sqlite_spec.clj
@@ -1,6 +1,6 @@
 (ns hyperion.sqlite-spec
   (:require [speclj.core :refer :all]
-            [hyperion.api :refer [*ds* new-datastore]]
+            [hyperion.api :refer [*ds* new-datastore defentity pack unpack]]
             [hyperion.log :as log]
             [hyperion.dev.spec :refer [it-behaves-like-a-datastore]]
             [hyperion.sql.connection :refer [with-connection]]
@@ -16,7 +16,7 @@
     (make-query query)))
 
 (def create-table-query
-  "CREATE TABLE %s (
+  "CREATE TABLE IF NOT EXISTS %s (
     id INTEGER PRIMARY KEY,
     name VARCHAR(35),
     first_name VARCHAR(35),
@@ -27,14 +27,14 @@
 
 (defn create-key-tables []
   (do-query
-    "CREATE TABLE account (
+    "CREATE TABLE IF NOT EXISTS account (
     id INTEGER PRIMARY KEY,
     first_name VARCHAR(35),
     inti INTEGER,
     data VARCHAR(32)
     );")
   (do-query
-    "CREATE TABLE shirt (
+    "CREATE TABLE IF NOT EXISTS shirt (
     id INTEGER PRIMARY KEY,
     account_id INTEGER,
     first_name VARCHAR(35),
@@ -42,6 +42,18 @@
     data VARCHAR(32),
     FOREIGN KEY (account_id) REFERENCES account(id)
     )"))
+
+(defn create-types-table []
+  (do-query
+    "CREATE TABLE IF NOT EXISTS types (
+    id INTEGER PRIMARY KEY,
+    inti INTEGER,
+    data VARCHAR(32),
+    bool BOOLEAN
+    )"))
+
+(defentity :types
+  [bool :type java.lang.Boolean])
 
 (defn create-table [table-name]
   (do-query (format create-table-query table-name)))
@@ -53,21 +65,24 @@
 
 (describe "SQLite Datastore"
   (around [it]
-    (binding [*ds* (new-datastore :implementation :sqlite :connection-url connection-url)]
+    (binding [*ds* (new-datastore :implementation :sqlite
+                                  :connection-url connection-url)]
       (it)))
 
   (before
     (with-connection connection-url
       (create-table "testing")
       (create-table "other_testing")
-      (create-key-tables)))
+      (create-key-tables)
+      (create-types-table)))
 
   (after
     (with-connection connection-url
       (drop-table "testing")
       (drop-table "other_testing")
       (drop-table "account")
-      (drop-table "shirt")))
+      (drop-table "shirt")
+      (drop-table "types")))
 
   (it-behaves-like-a-datastore)