+++ sqlite/src/hyperion/sqlite.clj
@@ -1,10 +1,11 @@
 (ns hyperion.sqlite
-  (:require [hyperion.sql.query :refer [add-to-query]]
+  (:require [chee.util :refer [->options]]
+            [hyperion.sorting :as sort]
+            [hyperion.sql.query :refer [add-to-query]]
             [hyperion.sql.query-builder :refer :all]
             [hyperion.sql.format :refer [column->db]]
             [hyperion.sql :refer :all]
-            [hyperion.sorting :as sort]
-            [chee.util :refer [->options]]))
+            [hyperion.sqlite.types :refer :all]))
 
 (clojure.lang.RT/loadClassForName "org.sqlite.JDBC")