+++ sql/src/hyperion/sql/spec_helper.clj
@@ -1,8 +1,8 @@
 (ns hyperion.sql.spec-helper
   (:require [speclj.core :refer :all]
-            [hyperion.sql.connection :refer [with-connection]]
-            [hyperion.sql.jdbc :refer [rollback]])
-  (:import [speclj SpecFailure]))
+            [speclj.platform]
+            [hyperion.sql.connection :refer [with-connection-url]]
+            [hyperion.sql.jdbc :refer [rollback]]))
 
 (defn str-contains? [s sub]
   (not= -1 (.indexOf s sub)))
@@ -10,14 +10,14 @@
 (defmacro error-msg-contains? [msg form]
   `(try
      ~form
-     (throw (SpecFailure. (str "Expected and exception with a message containing " ~msg " but no exception was thrown")))
+     (throw (-fail (str "Expected and exception with a message containing " ~msg " but no exception was thrown")))
      (catch Exception e#
        (let [message# (.getMessage e#)]
          (cond
-          (.isInstance SpecFailure e#)
+          (speclj.platform/failure? e#)
             (throw e#)
           (not (str-contains? (.getMessage e#) ~msg))
-            (throw (SpecFailure. (str "Expected and exception with a message containing \"" ~msg "\" but got: \"" message# "\"")))
+            (throw (-fail (str "Expected and exception with a message containing \"" ~msg "\" but got: \"" message# "\"")))
           :else e#)))))
 
 (defn with-rollback [url]